Clean up trash that gpt made for me
