+++ vue.config.js
@@ -1,4 +1,13 @@
-const { defineConfig } = require('@vue/cli-service')
+const { defineConfig } = require('@vue/cli-service');
+const path = require('path');
+
 module.exports = defineConfig({
-  transpileDependencies: true
-})
+  transpileDependencies: true,
+  configureWebpack: {
+    resolve: {
+      alias: {
+        '@utils': path.resolve(__dirname, 'src/utils')
+      }
+    }
+  }
+});
